pyproject.toml
[project]
name = "athena"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexis Beaulieu",email = "alexisbeaulieu97@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pluggy (>=1.5.0,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "dynaconf (>=3.2.10,<4.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
]


[tool.poetry.scripts]
athena = "athena.cli:app"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
